Hide Account and logout links from footer if user is not logged in
To be more precise, we would have to check also that the user is really
logged in to kilometrikisa too since now it is possible to only log in to
strava to get into the account page.

Perhaps that is better to be solved later if we decide to change the
login flow to be a bit more robust than the old one.